Clean up Volumes CMR (#7320)
