+++ packages/manager/src/features/Volumes/VolumeTableRow.test.tsx
@@ -1,8 +1,9 @@
-import { shallow } from 'enzyme';
+import { VolumeStatus } from '@linode/api-v4/lib/volumes';
+import { render } from '@testing-library/react';
 import * as React from 'react';
-
+import { wrapWithTheme } from 'src/utilities/testHelpers';
 import { volumes } from 'src/__data__/volumes';
-import { VolumeTableRow } from './VolumeTableRow';
+import { VolumeTableRow, CombinedProps } from './VolumeTableRow';
 
 const volumeWithLinodeLabel = {
   ...volumes[2],
@@ -12,25 +13,21 @@ const volumeWithLinodeLabel = {
 const unattachedVolume = {
   ...volumes[0],
   linodeLabel: '',
+  linode_id: null,
   linodeStatus: 'active'
 };
 
-const classes = {
-  root: '',
-  title: '',
-  labelCol: '',
-  labelStatusWrapper: '',
-  attachmentCol: '',
-  sizeCol: '',
-  pathCol: '',
-  volumesWrapper: '',
-  linodeVolumesWrapper: '',
-  systemPath: ''
-};
-
-const props = {
-  classes,
-  volume: volumeWithLinodeLabel,
+const props: CombinedProps = {
+  id: 0,
+  label: volumeWithLinodeLabel.linodeLabel,
+  region: '',
+  size: 0,
+  status: 'active' as VolumeStatus,
+  tags: [],
+  created: '',
+  updated: '',
+  filesystem_path: '',
+  linode_id: 0,
   isUpdating: false,
   isVolumesLanding: true,
   openForEdit: jest.fn(),
@@ -42,22 +39,17 @@ const props = {
   handleDelete: jest.fn()
 };
 
-const component = shallow(<VolumeTableRow {...props} />);
-
 describe('Volume table row', () => {
   it("should show the attached Linode's label if present", () => {
-    expect(
-      component
-        .find('[data-qa-volume-cell-attachment]')
-        .contains(volumeWithLinodeLabel.linodeLabel)
-    ).toBeTruthy();
+    const { getByText } = render(wrapWithTheme(<VolumeTableRow {...props} />));
+    expect(getByText(volumeWithLinodeLabel.linodeLabel));
   });
 
   it('should show Unattached if the Volume is not attached to a Linode', () => {
     const unattachedProps = { ...props, volume: unattachedVolume };
-    const unattached = shallow(<VolumeTableRow {...unattachedProps} />);
-    expect(
-      unattached.find('[data-qa-volume-cell-attachment]').contains('Unattached')
-    ).toBeTruthy();
+    const { getByText } = render(
+      wrapWithTheme(<VolumeTableRow {...unattachedProps} />)
+    );
+    expect(getByText('Detach'));
   });
 });
