+++ packages/manager/src/features/Volumes/VolumesLanding.tsx
@@ -4,29 +4,25 @@ import { Volume } from '@linode/api-v4/lib/volumes';
 import { withSnackbar, WithSnackbarProps } from 'notistack';
 import * as React from 'react';
 import { connect } from 'react-redux';
-import { RouteComponentProps, withRouter } from 'react-router-dom';
+import { RouteComponentProps } from 'react-router-dom';
 import { compose } from 'recompose';
 import { bindActionCreators, Dispatch } from 'redux';
-import VolumesIcon from 'src/assets/icons/entityIcons/volume.svg';
-import AddNewLink from 'src/components/AddNewLink';
-import Breadcrumb from 'src/components/Breadcrumb';
-import FormControlLabel from 'src/components/core/FormControlLabel';
-import {
-  createStyles,
-  Theme,
-  withStyles,
-  WithStyles
-} from 'src/components/core/styles';
+import VolumeIcon from 'src/assets/icons/entityIcons/volume.svg';
+import { makeStyles } from 'src/components/core/styles';
 import Typography from 'src/components/core/Typography';
-import setDocs from 'src/components/DocsSidebar/setDocs';
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
-import Grid from 'src/components/Grid';
-import OrderBy from 'src/components/OrderBy';
+import EntityTable_CMR from 'src/components/EntityTable/EntityTable_CMR';
+import LandingHeader from 'src/components/LandingHeader';
+import Loading from 'src/components/LandingLoading';
+import Link from 'src/components/Link';
 import { PaginationProps } from 'src/components/Pagey';
 import Placeholder from 'src/components/Placeholder';
-import Toggle from 'src/components/Toggle';
-import { REFRESH_INTERVAL } from 'src/constants';
+import PreferenceToggle from 'src/components/PreferenceToggle';
+import { ToggleProps } from 'src/components/PreferenceToggle/PreferenceToggle';
 import _withEvents, { EventsProps } from 'src/containers/events.container';
+import withRegions, {
+  DefaultProps as RegionProps
+} from 'src/containers/regions.container';
 import withVolumes, {
   StateProps as WithVolumesProps
 } from 'src/containers/volumes.container';
@@ -36,9 +32,8 @@ import withVolumesRequests, {
 import withLinodes, {
   Props as WithLinodesProps
 } from 'src/containers/withLinodes.container';
-import { BlockStorage } from 'src/documentation';
 import { resetEventsPolling } from 'src/eventsPolling';
-import LinodePermissionsError from 'src/features/linodes/LinodesDetail/LinodePermissionsError';
+import useReduxLoad from 'src/hooks/useReduxLoad';
 import {
   LinodeOptions,
   openForClone,
@@ -49,98 +44,11 @@ import {
   Origin as VolumeDrawerOrigin
 } from 'src/store/volumeForm';
 import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
-import { sendGroupByTagEnabledEvent } from 'src/utilities/ga';
 import DestructiveVolumeDialog from './DestructiveVolumeDialog';
-import ListGroupedVolumes from './ListGroupedVolumes';
-import ListVolumes from './ListVolumes';
-import VolumeAttachmentDrawer from './VolumeAttachmentDrawer';
-
-import ErrorState from 'src/components/ErrorState';
-import Loading from 'src/components/LandingLoading';
-import PreferenceToggle, { ToggleProps } from 'src/components/PreferenceToggle';
-
-import withRegions, {
-  DefaultProps as RegionProps
-} from 'src/containers/regions.container';
-import { doesRegionSupportBlockStorage } from 'src/utilities/doesRegionSupportBlockStorage';
 import { ExtendedVolume } from './types';
-
-type ClassNames =
-  | 'root'
-  | 'titleWrapper'
-  | 'title'
-  | 'tagGroup'
-  | 'labelCol'
-  | 'icon'
-  | 'attachmentCol'
-  | 'sizeCol'
-  | 'pathCol'
-  | 'volumesWrapper'
-  | 'linodeVolumesWrapper';
-
-const styles = (theme: Theme) =>
-  createStyles({
-    root: {
-      paddingBottom: 0
-    },
-    tagGroup: {
-      flexDirection: 'row-reverse',
-      position: 'relative',
-      top: -(theme.spacing(1) + 1)
-    },
-    titleWrapper: {
-      flex: 1
-    },
-    title: {
-      marginBottom: theme.spacing(1)
-    },
-    // styles for /volumes table
-    volumesWrapper: {},
-    // styles for linodes/id/volumes table
-    linodeVolumesWrapper: {
-      '& $labelCol': {
-        width: '20%',
-        minWidth: 200
-      },
-      '& $sizeCol': {
-        width: '15%',
-        minWidth: 100
-      },
-      '& $pathCol': {
-        width: '55%',
-        minWidth: 350
-      }
-    },
-    labelCol: {
-      width: '25%',
-      minWidth: 150,
-      paddingLeft: 65
-    },
-    icon: {
-      position: 'relative',
-      top: 3,
-      width: 40,
-      height: 40,
-      '& .circle': {
-        fill: theme.bg.offWhiteDT
-      },
-      '& .outerCircle': {
-        stroke: theme.bg.main
-      }
-    },
-    attachmentCol: {
-      width: '15%',
-      minWidth: 150
-    },
-    sizeCol: {
-      width: '10%',
-      minWidth: 75
-    },
-    pathCol: {
-      width: '25%',
-      minWidth: 250
-    }
-  });
+import VolumeAttachmentDrawer from './VolumeAttachmentDrawer';
+import { ActionHandlers as VolumeHandlers } from './VolumesActionMenu';
+import VolumeTableRow from './VolumeTableRow';
 
 interface Props {
   isVolumesLanding?: boolean;
@@ -182,24 +90,6 @@ interface DispatchProps {
   openForConfig: (volumeLabel: string, volumePath: string) => void;
 }
 
-interface State {
-  attachmentDrawer: {
-    open: boolean;
-    volumeId?: number;
-    volumeLabel?: string;
-    linodeRegion?: string;
-  };
-  destructiveDialog: {
-    open: boolean;
-    mode: 'detach' | 'delete';
-    volumeLabel: string;
-    volumeId?: number;
-    linodeLabel: string;
-    poweredOff?: boolean;
-    error?: string;
-  };
-}
-
 type RouteProps = RouteComponentProps<{ linodeId: string }>;
 
 type CombinedProps = Props &
@@ -212,385 +102,148 @@ type CombinedProps = Props &
   RouteProps &
   WithSnackbarProps &
   WithMappedVolumesProps &
-  WithStyles<ClassNames> &
   RegionProps;
 
-class VolumesLanding extends React.Component<CombinedProps, State> {
-  state: State = {
-    attachmentDrawer: {
+const volumeHeaders = [
+  {
+    label: 'Label',
+    dataColumn: 'label',
+    sortable: true,
+    widthPercent: 25
+  },
+  {
+    label: 'Region',
+    dataColumn: 'region',
+    sortable: true,
+    widthPercent: 15
+  },
+  {
+    label: 'Size',
+    dataColumn: 'size',
+    sortable: true,
+    widthPercent: 5
+  },
+  {
+    label: 'File System Path',
+    dataColumn: 'File System Path',
+    sortable: false,
+    widthPercent: 25,
+    hideOnMobile: true
+  },
+  {
+    label: 'Attached To',
+    dataColumn: 'Attached To',
+    sortable: false,
+    widthPercent: 20
+  }
+];
+
+const useStyles = makeStyles(() => ({
+  empty: {
+    '& svg': {
+      transform: 'scale(0.75)'
+    }
+  }
+}));
+
+export const VolumesLanding: React.FC<CombinedProps> = props => {
+  const classes = useStyles();
+
+  const {
+    volumesLoading,
+    mappedVolumesDataWithLinodes,
+    volumesLastUpdated,
+    volumesError,
+    openForConfig,
+    openForClone,
+    openForEdit,
+    openForResize
+  } = props;
+
+  const [attachmentDrawer, setAttachmentDrawer] = React.useState({
+    open: false,
+    volumeId: 0,
+    volumeLabel: '',
+    linodeRegion: ''
+  });
+
+  const [destructiveDialog, setDestructiveDialog] = React.useState<{
+    open: boolean;
+    mode: 'detach' | 'delete';
+    volumeId?: number;
+    volumeLabel: string;
+    linodeLabel: string;
+    error?: string;
+    poweredOff?: boolean;
+  }>({
+    open: false,
+    mode: 'detach',
+    volumeId: 0,
+    volumeLabel: '',
+    linodeLabel: '',
+    error: '',
+    poweredOff: false
+  });
+
+  const { _loading } = useReduxLoad(['volumes', 'linodes']);
+
+  const handleCloseAttachDrawer = () => {
+    setAttachmentDrawer(attachmentDrawer => ({
+      ...attachmentDrawer,
       open: false
-    },
-    destructiveDialog: {
-      open: false,
-      mode: 'detach',
-      volumeLabel: '',
-      linodeLabel: ''
-    }
+    }));
   };
 
-  mounted: boolean = false;
-
-  static docs: Linode.Doc[] = [
-    BlockStorage,
-    {
-      title: 'Boot a Linode from a Block Storage Volume',
-      src: `https://www.linode.com/docs/platform/block-storage/boot-from-block-storage-volume/`,
-      body: `This guide shows how to boot a Linode from a Block Storage Volume.`
-    }
-  ];
-
-  componentDidMount() {
-    const { getAllVolumes, volumesLastUpdated } = this.props;
-    this.mounted = true;
-    // If we haven't requested Volumes, or it's been a while, request them
-    if (Date.now() - volumesLastUpdated > REFRESH_INTERVAL) {
-      getAllVolumes().catch(_ => null); // Errors through Redux
-    }
-  }
-
-  componentWillUnmount() {
-    this.mounted = false;
-  }
-
-  handleCloseAttachDrawer = () => {
-    this.setState({ attachmentDrawer: { open: false } });
+  const handleAttach = (volumeId: number, label: string, regionID: string) => {
+    setAttachmentDrawer(attachmentDrawer => ({
+      ...attachmentDrawer,
+      open: true,
+      volumeId,
+      volumeLabel: label,
+      linodeRegion: regionID
+    }));
   };
 
-  handleAttach = (volumeId: number, label: string, regionID: string) => {
-    this.setState({
-      attachmentDrawer: {
-        open: true,
-        volumeId,
-        volumeLabel: label,
-        linodeRegion: regionID
-      }
-    });
-  };
-
-  handleDetach = (
+  const handleDetach = (
     volumeId: number,
     volumeLabel: string,
     linodeLabel: string,
     poweredOff: boolean
   ) => {
-    this.setState({
-      destructiveDialog: {
-        open: true,
-        mode: 'detach',
-        volumeId,
-        volumeLabel,
-        linodeLabel,
-        poweredOff,
-        error: undefined
-      }
-    });
+    setDestructiveDialog(destructiveDialog => ({
+      ...destructiveDialog,
+      open: true,
+      mode: 'detach',
+      volumeId,
+      volumeLabel,
+      linodeLabel,
+      poweredOff,
+      error: ''
+    }));
   };
 
-  handleDelete = (volumeId: number, volumeLabel: string) => {
-    this.setState({
-      destructiveDialog: {
-        open: true,
-        mode: 'delete',
-        volumeId,
-        volumeLabel,
-        linodeLabel: '',
-        error: undefined
-      }
-    });
+  const handleDelete = (volumeId: number, volumeLabel: string) => {
+    setDestructiveDialog(destructiveDialog => ({
+      ...destructiveDialog,
+      open: true,
+      mode: 'delete',
+      volumeId,
+      volumeLabel,
+      linodeLabel: '',
+      error: ''
+    }));
   };
 
-  render() {
-    const {
-      classes,
-      volumesError,
-      volumesLoading,
-      mappedVolumesDataWithLinodes,
-      readOnly,
-      removeBreadCrumb,
-      fromLinodes
-    } = this.props;
-
-    if (volumesLoading) {
-      return <Loading shouldDelay />;
-    }
-
-    if (volumesError && volumesError.read) {
-      return <RenderError />;
-    }
-
-    // If this is the Volumes tab on a Linode, we want ONLY the Volumes attached to this Linode.
-    const data =
-      mappedVolumesDataWithLinodes && this.props.linodeId
-        ? mappedVolumesDataWithLinodes.filter(
-            vol => vol.linode_id === this.props.linodeId
-          )
-        : mappedVolumesDataWithLinodes;
-
-    if (data.length < 1) {
-      return this.renderEmpty();
-    }
-
-    return (
-      <React.Fragment>
-        <DocumentTitleSegment segment="Volumes" />
-        {readOnly && <LinodePermissionsError />}
-        <PreferenceToggle<boolean>
-          preferenceKey="volumes_group_by_tag"
-          preferenceOptions={[false, true]}
-          localStorageKey="GROUP_VOLUMES"
-          toggleCallbackFnDebounced={toggleVolumesGroupBy}
-        >
-          {({
-            preference: volumesAreGrouped,
-            togglePreference: toggleGroupVolumes
-          }: ToggleProps<boolean>) => {
-            return (
-              <React.Fragment>
-                <Grid
-                  container
-                  justify="space-between"
-                  alignItems={removeBreadCrumb ? 'center' : 'flex-end'}
-                  className={classes.root}
-                >
-                  <Grid item className={classes.titleWrapper}>
-                    {removeBreadCrumb ? (
-                      <Typography variant="h2">Volumes</Typography>
-                    ) : (
-                      <Breadcrumb
-                        pathname={this.props.location.pathname}
-                        labelTitle="Volumes"
-                        className={classes.title}
-                      />
-                    )}
-                  </Grid>
-                  <Grid item className="p0">
-                    <FormControlLabel
-                      className={classes.tagGroup}
-                      control={
-                        <Toggle
-                          className={
-                            volumesAreGrouped ? ' checked' : ' unchecked'
-                          }
-                          onChange={toggleGroupVolumes}
-                          checked={volumesAreGrouped}
-                        />
-                      }
-                      label="Group by Tag:"
-                    />
-                  </Grid>
-                  <Grid item>
-                    <Grid container alignItems="flex-end">
-                      <Grid item className="pt0">
-                        <AddNewLink
-                          onClick={
-                            fromLinodes
-                              ? this.openCreateVolumeDrawer
-                              : () => {
-                                  this.props.history.push('/volumes/create');
-                                }
-                          }
-                          label="Create a Volume"
-                        />
-                      </Grid>
-                    </Grid>
-                  </Grid>
-                </Grid>
-                {this.renderData(data, volumesAreGrouped)}
-              </React.Fragment>
-            );
-          }}
-        </PreferenceToggle>
-
-        <VolumeAttachmentDrawer
-          open={this.state.attachmentDrawer.open}
-          volumeId={this.state.attachmentDrawer.volumeId || 0}
-          volumeLabel={this.state.attachmentDrawer.volumeLabel || ''}
-          linodeRegion={this.state.attachmentDrawer.linodeRegion || ''}
-          onClose={this.handleCloseAttachDrawer}
-        />
-        <DestructiveVolumeDialog
-          open={this.state.destructiveDialog.open}
-          error={this.state.destructiveDialog.error}
-          volumeLabel={this.state.destructiveDialog.volumeLabel}
-          linodeLabel={this.state.destructiveDialog.linodeLabel}
-          poweredOff={this.state.destructiveDialog.poweredOff || false}
-          mode={this.state.destructiveDialog.mode}
-          onClose={this.closeDestructiveDialog}
-          onDetach={this.detachVolume}
-          onDelete={this.deleteVolume}
-        />
-      </React.Fragment>
-    );
-  }
-
-  goToSettings = () => {
-    const { history, linodeId } = this.props;
-    history.push(`/linodes/${linodeId}/settings`);
+  const closeDestructiveDialog = () => {
+    setDestructiveDialog(destructiveDialog => ({
+      ...destructiveDialog,
+      open: false
+    }));
   };
 
-  renderEmpty = () => {
-    const {
-      isVolumesLanding,
-      linodeConfigs,
-      linodeRegion,
-      readOnly,
-      regionsData,
-      fromLinodes
-    } = this.props;
-
-    if (
-      linodeRegion &&
-      !doesRegionSupportBlockStorage(linodeRegion, regionsData)
-    ) {
-      return (
-        <React.Fragment>
-          <DocumentTitleSegment segment="Volumes" />
-          <Placeholder
-            title="Volumes are not available in this region"
-            isEntity
-            icon={VolumesIcon}
-            renderAsSecondary={!isVolumesLanding}
-          />
-        </React.Fragment>
-      );
-    }
-
-    if (linodeConfigs && linodeConfigs.length === 0) {
-      return (
-        <React.Fragment>
-          <DocumentTitleSegment segment="Volumes" />
-          <Placeholder
-            title="No configs available."
-            isEntity
-            icon={VolumesIcon}
-            buttonProps={[
-              {
-                onClick: this.goToSettings,
-                children: 'View Linode Configurations'
-              }
-            ]}
-            renderAsSecondary={!isVolumesLanding}
-          >
-            This Linode has no configurations. Click below to create a
-            configuration.
-          </Placeholder>
-        </React.Fragment>
-      );
-    }
-
-    return (
-      <React.Fragment>
-        <DocumentTitleSegment segment="Volumes" />
-        {readOnly && <LinodePermissionsError />}
-        <Placeholder
-          title="Add Block Storage!"
-          isEntity
-          icon={VolumesIcon}
-          renderAsSecondary={!isVolumesLanding}
-          buttonProps={[
-            {
-              onClick: fromLinodes
-                ? this.openCreateVolumeDrawer
-                : () => {
-                    this.props.history.push('/volumes/create');
-                  },
-
-              children: 'Add a Volume',
-              disabled: readOnly
-            }
-          ]}
-        >
-          <Typography variant="subtitle1">Need additional storage?</Typography>
-          <Typography variant="subtitle1">
-            <a
-              href="https://linode.com/docs/platform/block-storage/how-to-use-block-storage-with-your-linode-new-manager/"
-              target="_blank"
-              aria-describedby="external-site"
-              rel="noopener noreferrer"
-              className="h-u"
-            >
-              Here&apos;s how to use Block Storage with your Linode
-            </a>
-            &nbsp;or&nbsp;
-            <a
-              href="https://www.linode.com/docs/"
-              target="_blank"
-              aria-describedby="external-site"
-              rel="noopener noreferrer"
-              className="h-u"
-            >
-              visit our guides and tutorials.
-            </a>
-          </Typography>
-        </Placeholder>
-      </React.Fragment>
-    );
-  };
-
-  renderData = (volumes: ExtendedVolume[], volumesAreGrouped: boolean) => {
-    const renderProps = {
-      isVolumesLanding: Boolean(this.props.isVolumesLanding),
-      handleAttach: this.handleAttach,
-      handleDelete: this.handleDelete,
-      handleDetach: this.handleDetach,
-      openForEdit: this.props.openForEdit,
-      openForClone: this.props.openForClone,
-      openForConfig: this.props.openForConfig,
-      openForResize: this.props.openForResize
-    };
-
-    return (
-      <OrderBy data={volumes} order={'asc'} orderBy={'label'}>
-        {({ data: orderedData, handleOrderChange, order, orderBy }) => {
-          const orderProps = {
-            orderBy,
-            order,
-            handleOrderChange,
-            data: orderedData
-          };
-
-          return volumesAreGrouped ? (
-            <ListGroupedVolumes
-              {...orderProps}
-              renderProps={{ ...renderProps }}
-            />
-          ) : (
-            <ListVolumes {...orderProps} renderProps={{ ...renderProps }} />
-          );
-        }}
-      </OrderBy>
-    );
-  };
-
-  closeDestructiveDialog = () => {
-    this.setState({
-      destructiveDialog: {
-        ...this.state.destructiveDialog,
-        open: false
-      }
-    });
-  };
-
-  openCreateVolumeDrawer = (e: any) => {
-    const { linodeId, linodeLabel, linodeRegion } = this.props;
-    if (linodeId && linodeLabel && linodeRegion) {
-      return this.props.openForCreating('Created from Linode Details', {
-        linodeId,
-        linodeLabel,
-        linodeRegion
-      });
-    }
-
-    this.props.openForCreating('Created from Volumes Landing');
-
-    e.preventDefault();
-  };
-
-  detachVolume = () => {
-    const {
-      destructiveDialog: { volumeId }
-    } = this.state;
-    const { detachVolume } = this.props;
+  const detachVolume = () => {
+    const { volumeId } = destructiveDialog;
+    const { detachVolume } = props;
     if (!volumeId) {
       return;
     }
@@ -598,28 +251,24 @@ class VolumesLanding extends React.Component<CombinedProps, State> {
     detachVolume({ volumeId })
       .then(_ => {
         /* @todo: show a progress bar for volume detachment */
-        this.props.enqueueSnackbar('Volume detachment started', {
+        props.enqueueSnackbar('Volume detachment started', {
           variant: 'info'
         });
-        this.closeDestructiveDialog();
+        closeDestructiveDialog();
         resetEventsPolling();
       })
       .catch(error => {
-        this.setState({
-          destructiveDialog: {
-            ...this.state.destructiveDialog,
-            error: getAPIErrorOrDefault(error, 'Unable to detach Volume.')[0]
-              .reason
-          }
-        });
+        setDestructiveDialog(destructiveDialog => ({
+          ...destructiveDialog,
+          error: getAPIErrorOrDefault(error, 'Unable to detach Volume.')[0]
+            .reason
+        }));
       });
   };
 
-  deleteVolume = () => {
-    const {
-      destructiveDialog: { volumeId }
-    } = this.state;
-    const { deleteVolume } = this.props;
+  const deleteVolume = () => {
+    const { volumeId } = destructiveDialog;
+    const { deleteVolume } = props;
 
     if (!volumeId) {
       return;
@@ -627,25 +276,127 @@ class VolumesLanding extends React.Component<CombinedProps, State> {
 
     deleteVolume({ volumeId })
       .then(() => {
-        this.closeDestructiveDialog();
+        closeDestructiveDialog();
         resetEventsPolling();
       })
       .catch(error => {
-        this.setState({
-          destructiveDialog: {
-            ...this.state.destructiveDialog,
-            error: getAPIErrorOrDefault(error, 'Unable to delete Volume.')[0]
-              .reason
-          }
-        });
+        setDestructiveDialog(destructiveDialog => ({
+          ...destructiveDialog,
+          error: getAPIErrorOrDefault(error, 'Unable to delete Volume.')[0]
+            .reason
+        }));
       });
   };
-}
 
-const eventCategory = `volumes landing`;
+  if (_loading) {
+    return <Loading />;
+  }
 
-const toggleVolumesGroupBy = (checked: boolean) =>
-  sendGroupByTagEnabledEvent(eventCategory, checked);
+  if (
+    mappedVolumesDataWithLinodes.length === 0 &&
+    !volumesError.read &&
+    volumesLastUpdated > 0
+  ) {
+    return (
+      <>
+        <DocumentTitleSegment segment="Volumes" />
+        <Placeholder
+          title="Volumes"
+          className={classes.empty}
+          icon={VolumeIcon}
+          isEntity
+          buttonProps={[
+            {
+              onClick: () => props.history.push('/volumes/create'),
+              children: 'Add a Volume'
+            }
+          ]}
+        >
+          <Typography variant="subtitle1">
+            Attach additional storage to your Linode.
+          </Typography>
+          <Typography variant="subtitle1">
+            <Link to="https://www.linode.com/docs/products/storage/block-storage/">
+              Learn more about Linode Block Storage Volumes.
+            </Link>
+          </Typography>
+        </Placeholder>
+      </>
+    );
+  }
+
+  const handlers: VolumeHandlers = {
+    openForConfig,
+    openForEdit,
+    openForResize,
+    openForClone,
+    handleAttach,
+    handleDetach,
+    handleDelete
+  };
+
+  const volumeRow = {
+    handlers,
+    Component: VolumeTableRow,
+    data: mappedVolumesDataWithLinodes ?? [],
+    loading: volumesLoading,
+    lastUpdated: volumesLastUpdated,
+    error: volumesError.read
+  };
+
+  return (
+    <>
+      <DocumentTitleSegment segment="Volumes" />
+      <PreferenceToggle<boolean>
+        preferenceKey="volumes_group_by_tag"
+        preferenceOptions={[false, true]}
+        localStorageKey="GROUP_VOLUMES"
+      >
+        {({
+          preference: volumesAreGrouped,
+          togglePreference: toggleGroupVolumes
+        }: ToggleProps<boolean>) => {
+          return (
+            <>
+              <LandingHeader
+                title="Volumes"
+                entity="Volume"
+                onAddNew={() => props.history.push('/volumes/create')}
+                docsLink="https://www.linode.com/docs/platform/block-storage/how-to-use-block-storage-with-your-linode/"
+              />
+              <EntityTable_CMR
+                entity="volume"
+                headers={volumeHeaders}
+                isGroupedByTag={volumesAreGrouped}
+                toggleGroupByTag={toggleGroupVolumes}
+                row={volumeRow}
+                initialOrder={{ order: 'asc', orderBy: 'label' }}
+              />
+              <VolumeAttachmentDrawer
+                open={attachmentDrawer.open}
+                volumeId={attachmentDrawer.volumeId || 0}
+                volumeLabel={attachmentDrawer.volumeLabel || ''}
+                linodeRegion={attachmentDrawer.linodeRegion || ''}
+                onClose={handleCloseAttachDrawer}
+              />
+              <DestructiveVolumeDialog
+                open={destructiveDialog.open}
+                error={destructiveDialog.error}
+                volumeLabel={destructiveDialog.volumeLabel}
+                linodeLabel={destructiveDialog.linodeLabel}
+                poweredOff={destructiveDialog.poweredOff || false}
+                mode={destructiveDialog.mode}
+                onClose={closeDestructiveDialog}
+                onDetach={detachVolume}
+                onDelete={deleteVolume}
+              />
+            </>
+          );
+        }}
+      </PreferenceToggle>
+    </>
+  );
+};
 
 const mapDispatchToProps = (dispatch: Dispatch) =>
   bindActionCreators(
@@ -660,10 +411,6 @@ const mapDispatchToProps = (dispatch: Dispatch) =>
   );
 
 const connected = connect(undefined, mapDispatchToProps);
-
-const documented = setDocs(VolumesLanding.docs);
-
-const styled = withStyles(styles);
 
 const addAttachedLinodeInfoToVolume = (
   volume: Volume,
@@ -703,9 +450,7 @@ const filterVolumeEvents = (event: Event): boolean => {
 
 export default compose<CombinedProps, Props>(
   connected,
-  documented,
   withVolumesRequests,
-  withRouter,
   _withEvents((ownProps: CombinedProps, eventsData) => ({
     ...ownProps,
     eventsData: eventsData.filter(filterVolumeEvents)
@@ -741,12 +486,5 @@ export default compose<CombinedProps, Props>(
     }
   ),
   withRegions(),
-  withSnackbar,
-  styled
+  withSnackbar
 )(VolumesLanding);
-
-const RenderError = () => {
-  return (
-    <ErrorState errorText="There was an error loading your Volumes. Please try again later" />
-  );
-};
